requirement.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
docker==5.0.0
english==2020.7.0
filelock==3.0.12
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isodate==0.6.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
joblib==1.0.1
ltp==4.1.4
lxml==4.6.3
monotonic==1.6
neotime==1.7.4
numpy==1.20.2
packaging==20.9
pansi==2020.7.3
parsel==1.6.0
pdfminer.six==20200517
pdfplumber==0.5.27
Pillow==8.2.0
priority==1.3.0
prompt-toolkit==3.0.18
Protego==0.1.16
py2neo==2021.1.1
pyahocorasick==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyDispatcher==2.0.5
Pygments==2.9.0
pygtrie==2.4.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
queuelib==1.6.1
rdflib==5.0.0
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
Scrapy==2.5.0
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2.1
tokenizers==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.5.1
Twisted==21.2.0
typing-extensions==3.10.0.0
urllib3==1.26.4
w3lib==1.22.0
Wand==0.6.6
wcwidth==0.2.5
websocket-client==0.59.0
zope.interface==5.4.0
